backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
langchain = "^0.3.7"
faiss-cpu = "^1.9.0"
python-dotenv = "^1.0.1"
langsmith = "^0.1.142"
langchain-community = "^0.3.5"
langchain-huggingface = "^0.1.2"
huggingface-hub = "^0.26.2"
pydantic = "^2.9.2"
openai = "^1.54.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
